workers/gpu-worker/requirements.txt
# =============================================================================
# lore-anchor GPU Worker - Python Dependencies
# =============================================================================

# --- Deep Learning ---
torch==2.1.2+cu121
torchvision==0.16.2+cu121
--extra-index-url https://download.pytorch.org/whl/cu121

# --- Image Processing ---
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0

# --- Adversarial / Mist v2 ---
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0

# --- Queue / Task ---
redis>=5.0.0
arq>=0.25.0

# --- Cloud Storage (R2 via S3 API) ---
boto3>=1.34.0

# --- C2PA Signing ---
c2pa-python>=0.4.0

# --- Utilities ---
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
